[ci] Move datasource accessors into their own package
