+++ core/src/main/java/psiprobe/beans/accessors/DatasourceAccessor.java
@@ -8,7 +8,7 @@
  * WITHOUT LIMITATION, THE IMPLIED WARRANTIES OF MERCHANTIBILITY AND FITNESS FOR A PARTICULAR
  * PURPOSE.
  */
-package psiprobe.beans;
+package psiprobe.beans.accessors;
 
 import psiprobe.model.DataSourceInfo;
