+++ core/src/main/java/psiprobe/beans/accessors/TomcatJdbcPoolDatasourceAccessor.java
@@ -8,30 +8,30 @@
  * WITHOUT LIMITATION, THE IMPLIED WARRANTIES OF MERCHANTIBILITY AND FITNESS FOR A PARTICULAR
  * PURPOSE.
  */
-package psiprobe.beans;
+package psiprobe.beans.accessors;
 
-import org.apache.tomcat.dbcp.dbcp2.BasicDataSource;
+import org.apache.tomcat.jdbc.pool.DataSource;
 
 import psiprobe.model.DataSourceInfo;
 
 /**
- * The Class Tomcat8DbcpDatasourceAccessor.
+ * Datasource accessor for tomcat.
  */
-public class Tomcat8DbcpDatasourceAccessor implements DatasourceAccessor {
+public class TomcatJdbcPoolDatasourceAccessor implements DatasourceAccessor {
 
   @Override
   public DataSourceInfo getInfo(Object resource) throws Exception {
     DataSourceInfo dataSourceInfo = null;
     if (canMap(resource)) {
-      BasicDataSource source = (BasicDataSource) resource;
+      DataSource source = (DataSource) resource;
       dataSourceInfo = new DataSourceInfo();
       dataSourceInfo.setBusyConnections(source.getNumActive());
       dataSourceInfo.setEstablishedConnections(source.getNumIdle() + source.getNumActive());
-      dataSourceInfo.setMaxConnections(source.getMaxTotal());
+      dataSourceInfo.setMaxConnections(source.getMaxActive());
       dataSourceInfo.setJdbcUrl(source.getUrl());
       dataSourceInfo.setUsername(source.getUsername());
       dataSourceInfo.setResettable(false);
-      dataSourceInfo.setType("tomcat-dbcp2");
+      dataSourceInfo.setType("tomcat-jdbc");
     }
     return dataSourceInfo;
   }
@@ -43,8 +43,8 @@ public class Tomcat8DbcpDatasourceAccessor implements DatasourceAccessor {
 
   @Override
   public boolean canMap(Object resource) {
-    return "org.apache.tomcat.dbcp.dbcp2.BasicDataSource".equals(resource.getClass().getName())
-        && resource instanceof BasicDataSource;
+    return "org.apache.tomcat.jdbc.pool.DataSource".equals(resource.getClass().getName())
+        && resource instanceof DataSource;
   }
 
 }
