+++ core/src/main/java/psiprobe/beans/accessors/ViburCpDatasourceAccessor.java
@@ -8,7 +8,7 @@
  * WITHOUT LIMITATION, THE IMPLIED WARRANTIES OF MERCHANTIBILITY AND FITNESS FOR A PARTICULAR
  * PURPOSE.
  */
-package psiprobe.beans;
+package psiprobe.beans.accessors;
 
 import java.lang.management.ManagementFactory;
