+++ core/src/test/java/psiprobe/beans/accessors/OpenEjbManagedDatasourceAccessorTest.java
@@ -8,50 +8,49 @@
  * WITHOUT LIMITATION, THE IMPLIED WARRANTIES OF MERCHANTIBILITY AND FITNESS FOR A PARTICULAR
  * PURPOSE.
  */
-package psiprobe.beans;
+package psiprobe.beans.accessors;
 
-import java.sql.SQLException;
-import java.util.Properties;
-
+import org.apache.openejb.resource.jdbc.managed.local.ManagedDataSource;
 import org.junit.Assert;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import com.mchange.v2.c3p0.ComboPooledDataSource;
 
-import mockit.Expectations;
 import mockit.Mocked;
-import oracle.jdbc.pool.OracleDataSource;
+import mockit.Tested;
+import psiprobe.beans.accessors.OpenEjbManagedDatasourceAccessor;
 
 /**
- * The Class OracleDatasourceAccessorTest.
+ * The Class OpenEjbManagedDatasourceAccessorTest.
  */
-public class OracleDatasourceAccessorTest {
+public class OpenEjbManagedDatasourceAccessorTest {
 
     /** The accessor. */
-    OracleDatasourceAccessor accessor;
+    @Tested
+    OpenEjbManagedDatasourceAccessor accessor;
 
     /** The source. */
     @Mocked
-    OracleDataSource source;
+    ManagedDataSource source;
 
     /** The bad source. */
     ComboPooledDataSource badSource;
 
     /**
      * Before.
-     *
-     * @throws SQLException the SQL exception
      */
     @Before
-    public void before() throws SQLException {
-        accessor = new OracleDatasourceAccessor();
+    public void before() {
+        accessor = new OpenEjbManagedDatasourceAccessor();
         badSource = new ComboPooledDataSource();
     }
 
     /**
      * Can map test.
      */
+    @Ignore
     @Test
     public void canMapTest() {
         Assert.assertTrue(accessor.canMap(source));
@@ -73,12 +72,6 @@ public class OracleDatasourceAccessorTest {
      */
     @Test
     public void getInfoTest() throws Exception {
-        new Expectations() {
-            {
-                source.getConnectionCacheProperties();
-                result = new Properties();
-            }
-        };
         accessor.getInfo(source);
     }
