+++ tomcat80adapter/src/main/java/psiprobe/beans/accessors/package-info.java
@@ -11,4 +11,4 @@
 /**
  * Psi-Probe Beans Package.
  */
-package psiprobe.beans;
+package psiprobe.beans.accessors;
